BZ #1508408 - DB upgrade[6.2 -> 6.3] failed at db_migrate step: PG::Error: ERROR: update or delete on table "mail_notifications" violates foreign key constraint
Fix undefined method 'destroy' for nil

+++ db/migrate/20161220092005_remove_satellite_from_notification_name_spell_correction.rb
@@ -19,6 +19,6 @@ class RemoveSatelliteFromNotificationNameSpellCorrection < ActiveRecord::Migrati
       end
     end
 
-    satellite_sync_errata_notification.destroy
+    satellite_sync_errata_notification.destroy if satellite_sync_errata_notification.present?
   end
 end
